streamlit-app-template/requirements.txt
streamlit>=1.35
pandas>=2.0
Pillow>=9.5
databricks-sdk>=0.30
databricks-sql-connector>=2.9


